Add object storage interfaces and file system implementation
These interfaces are mainly supposed to be implemented with S3, GCS, and other object storage implementations, which are to be done. The file system implementation is mostly for testing.

To put this in the future context, the plugin will instantiate concrete implementations of `ObjectStorageFactory`, which will be S3, GCS, and others.