Identify encryption key in manifest
This commit is a preparation for changing how the plugin deals with encryption keys. The idea will be that there will be several RSA key encryption keys (KEK) provided in the configuration. When a data encryption key (DEK) is encrypted and written to the manifest, it comes with the ID of the KEK that encrypted it. When it's time to decrypt, the corresponding KEK is found by ID in the keyring and used for decryption.

This commit doesn't yet introduce multiple keys on the configuration level, but set's up all the necessary internal mechanics and uses a static KEK ID for now.

There will be a follow-up that actually introduces the keyring on the configuration level.