fix: add empty indexes to builder
By ignoring empty indexes--that is a valid case when segment sizes are too small; even if not recommended--the requirement to have at least 4 indexes when uploaded was not met.

This fix improves the logging to show what indexes are included therefore know which one is missing; and includes a fix to add empty indexes to the builder.