fix: add retention-based eviction to disk-based cache metrics test
To reduce flakiness where deletion is not executed in a consistent manner based on size, a time-based eviction configuration is added to have either 1 or 2 deletions happening while test is running to validate results.

Before it was only checking either first or second value where deleted. Now it is checking 1 or 2 or both.

To validated flakiness, @RepeatedTest(100000) was used, and it's now passing fine.